Refactor GameSquareDOM to functional component
